refactore: change structure from nestjs
