+++ package.json
@@ -37,7 +37,7 @@
     "@nestjs/schematics": "^10.0.0",
     "@nestjs/testing": "^10.0.0",
     "@swc/core": "^1.7.26",
-    "@types/express": "^4.17.17",
+    "@types/express": "^4.17.21",
     "@types/jest": "^29.5.13",
     "@types/node": "^20.3.1",
     "@types/supertest": "^6.0.0",
